+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-## [1.4.8] - 2026-03-31
+## [1.4.8] - 2026-04-02
 
 ### ✨ Added